Initial version
Jupyter notebook which illustrates Signal Detection Theory. 

In situations where a signal may be present among noise, an observer may decide a signal is present when a certain criterion is exceeded. In such a scenario, there are four possible outcomes: 
1. There is only noise, and the observer correctly decides there is none. This is called a Correct Rejection
2. There is only noise, but the observer erroneously decides there is one. This is called a False Alarm
3. There is a signal, but the observer decides there is none. This is called a Miss
4. There is a signal, and the observer decides there is one. This is called a hit.
Depending on the observer's criterion (and the discriminability of the signal from noise) the ratio of the four outcomes varies. This notebook features a slider which controls the value of the criterion, and allows the user of the script to visualise how the ratio of the outcomes changes for a given noise and signal distribution.